+++ quickstart/eabm_package/single_element_mesh.py
@@ -3,7 +3,6 @@ import os
 import argparse
 import inspect
 import shutil
-import re
 
 import abaqus
 import abaqusConstants