+++ waves/main.py
@@ -34,7 +34,8 @@ def main():
     elif args.subcommand == 'visualize':
         return_code = visualization(target=args.TARGET, output_file=args.output_file,
                                     sconstruct=args.sconstruct, print_graphml=args.print_graphml,
-                                    exclude_list=args.exclude_list, height=args.height, width=args.width)
+                                    exclude_list=args.exclude_list, exclude_regex=args.exclude_regex, 
+                                    height=args.height, width=args.width)
     else:
         parser.print_help()
 
@@ -157,6 +158,8 @@ def get_parser():
         help="Width of visualization in inches if being saved to a file (default: %(default)s)")
     visualize_parser.add_argument("-e", "--exclude-list", nargs="*", default=_settings._visualize_exclude,
         help="If a node starts or ends with one of these string literals, do not visualize it (default: %(default)s)")
+    visualize_parser.add_argument("-r", "--exclude-regex", type=str,
+        help="If a node matches this regular expression, do not visualize it (default: %(default)s)")
     visualize_parser.add_argument("-g", "--print-graphml", dest="print_graphml", action="store_true",
         help="Print the visualization in graphml format (default: %(default)s)")
 
@@ -179,7 +182,7 @@ def docs(print_local_path=False):
     return 0
 
 
-def build(targets, scons_args=[], max_iterations=5, working_directory=None, git_clone_directory=None):
+def build(targets, scons_args=None, max_iterations=5, working_directory=None, git_clone_directory=None):
     """Submit an iterative SCons command
 
     SCons command is re-submitted until SCons reports that the target 'is up to date.' or the iteration count is
@@ -190,6 +193,8 @@ def build(targets, scons_args=[], max_iterations=5, working_directory=None, git_
     :param int max_iterations: maximum number of iterations before the iterative loop is terminated
     :param str working_directory: Change the SCons command working directory
     """
+    if not scons_args:
+        scons_args = []
     if not targets:
         print("At least one target must be provided", file=sys.stderr)
         return 1
@@ -218,7 +223,7 @@ def build(targets, scons_args=[], max_iterations=5, working_directory=None, git_
     return 0
 
 
-def fetch(subcommand, root_directory, relative_paths, destination, requested_paths=[],
+def fetch(subcommand, root_directory, relative_paths, destination, requested_paths=None,
           overwrite=False, dry_run=False, print_available=False):
     """Thin wrapper on :meth:`waves.fetch.recursive_copy` to provide subcommand specific behavior and STDOUT/STDERR
 
@@ -238,6 +243,8 @@ def fetch(subcommand, root_directory, relative_paths, destination, requested_pat
     :param bool dry_run: Print the destination tree and exit. Short circuited by ``print_available``
     :param bool print_available: Print the available source files and exit. Short circuits ``dry_run``
     """
+    if not requested_paths:
+        requested_paths = []
     if not root_directory.is_dir():
         # During "waves quickstart/fetch" sub-command(s), this should only be reached if the package installation structure
         # doesn't match the assumptions in _settings.py. It is used by the Conda build tests as a sign-of-life that the
@@ -252,7 +259,7 @@ def fetch(subcommand, root_directory, relative_paths, destination, requested_pat
     return return_code
 
 
-def visualization(target, sconstruct, exclude_list, output_file=None, print_graphml=False,
+def visualization(target, sconstruct, exclude_list, exclude_regex, output_file=None, print_graphml=False,
                   height=_settings._visualize_default_height, width=_settings._visualize_default_width):
     """Visualize the directed acyclic graph created by a SCons build
 
@@ -263,6 +270,7 @@ def visualization(target, sconstruct, exclude_list, output_file=None, print_grap
     :param str target: String specifying an SCons target
     :param str sconstruct: Path to an SConstruct file or parent directory
     :param list exclude_list: exclude nodes starting with strings in this list (e.g. /usr/bin)
+    :param str exclude_regex: exclude nodes that match this regular expression
     :param str output_file: File for saving the visualization
     :param bool print_graphml: Whether to print the graph in graphml format
     :param int height: Height of visualization if being saved to a file
@@ -279,7 +287,7 @@ def visualization(target, sconstruct, exclude_list, output_file=None, print_grap
     scons_command.extend(_settings._scons_visualize_arguments)
     scons_stdout = subprocess.check_output(scons_command, cwd=sconstruct.parent)
     tree_output = scons_stdout.decode("utf-8").split('\n')
-    tree_dict = visualize.parse_output(tree_output, exclude_list=exclude_list)
+    tree_dict = visualize.parse_output(tree_output, exclude_list=exclude_list, exclude_regex=exclude_regex)
 
     if print_graphml:
         print(tree_dict['graphml'], file=sys.stdout)