+++ waves/parameter_generators.py
@@ -5,7 +5,6 @@ import sys
 import itertools
 import copy
 import hashlib
-import numbers
 import warnings
 
 import yaml
@@ -162,13 +161,15 @@ class _ParameterGenerator(ABC):
         if self.previous_parameter_study:
             self._merge_parameter_studies()
 
-    def generate(self, kwargs={}):
+    def generate(self, kwargs=None):
         """Deprecated public generate method.
 
         The parameter study is now generated as part of class instantiation. This method has been kept for backward
         compatibility. Method will overwrite the class instantiated study with a new parameter study each time it is
         called instead of duplicating or merging the parameter study.
         """
+        if not kwargs:
+            kwargs = {}
         warning_message = "Parameter studies are now generated during class instantiation. " \
                           "The generate method is deprecated and will be removed in a future release."
         warnings.warn(warning_message, DeprecationWarning)
@@ -216,7 +217,7 @@ class _ParameterGenerator(ABC):
         self.write()
 
     def _write_dataset(self):
-        """Write Xarray Datset formatted output to STDOUT, separate set files, or a single file
+        """Write Xarray Dataset formatted output to STDOUT, separate set files, or a single file
 
         Behavior as specified in :meth:`waves.parameter_generators._ParameterGenerator.write`
         """
@@ -908,11 +909,11 @@ class SobolSequence(_ScipyGenerator):
         super()._generate(**kwargs)
 
     def parameter_study_to_dict(self, *args, **kwargs):
-        # Get the ABC docstring into each paramter generator API
+        # Get the ABC docstring into each parameter generator API
         return super().parameter_study_to_dict(*args, **kwargs)
 
     def write(self):
-        # Get the ABC docstring into each paramter generator API
+        # Get the ABC docstring into each parameter generator API
         super().write()
 
 
@@ -1004,11 +1005,11 @@ class ScipySampler(_ScipyGenerator):
         super()._generate(**kwargs)
 
     def parameter_study_to_dict(self, *args, **kwargs):
-        # Get the ABC docstring into each paramter generator API
+        # Get the ABC docstring into each parameter generator API
         return super().parameter_study_to_dict(*args, **kwargs)
 
     def write(self):
-        # Get the ABC docstring into each paramter generator API
+        # Get the ABC docstring into each parameter generator API
         super().write()
 
 
@@ -1144,7 +1145,7 @@ class SALibSampler(_ParameterGenerator, ABC):
         if self.sampler_class == "morris" and parameter_count < 2:
             raise ValueError("The SALib Morris sampler requires at least two parameters")
 
-    def _sampler_overrides(self, override_kwargs={}):
+    def _sampler_overrides(self, override_kwargs=None):
         """Provide sampler specific kwarg override dictionaries
 
         * sobol produces duplicate parameter sets for two parameters when ``calc_second_order`` is ``True``. Override
@@ -1154,6 +1155,8 @@ class SALibSampler(_ParameterGenerator, ABC):
         :return: override kwarg dictionary
         :rtype: dict
         """
+        if not override_kwargs:
+            override_kwargs = {}
         parameter_count = len(self._parameter_names)
         if self.sampler_class == "sobol" and parameter_count == 2:
             override_kwargs = {**override_kwargs, "calc_second_order": False}
@@ -1181,9 +1184,9 @@ class SALibSampler(_ParameterGenerator, ABC):
         super()._generate()
 
     def parameter_study_to_dict(self, *args, **kwargs):
-        # Get the ABC docstring into each paramter generator API
+        # Get the ABC docstring into each parameter generator API
         return super().parameter_study_to_dict(*args, **kwargs)
 
     def write(self):
-        # Get the ABC docstring into each paramter generator API
+        # Get the ABC docstring into each parameter generator API
         super().write()