+++ waves/tests/test_scipy_sampler.py
@@ -2,8 +2,6 @@
 """
 
 from unittest.mock import patch
-from contextlib import nullcontext as does_not_raise
-import inspect
 
 import pytest
 import numpy